pyproject.toml
[tool.poetry]
name = "simplesoilprofile"
version = "0.0.1"
description = "Simple soil profile for hydrological modelling (particularily SWAP)."
authors = ["Mateusz Zawadzki <zawadzkimat2@gmail.com>"]
repository = "https://github.com/zawadzkim/simplesoilprofile"
documentation = "https://zawadzkim.github.io/simplesoilprofile/"
readme = "README.md"
packages = [
  {include = "simplesoilprofile"}
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pydantic = "^2.4.2"
numpy = "^1.24.0"
matplotlib = "^3.8.0"
pyyaml = "^6.0.1"
requests = "^2.31.0"
jsonpath-ng = "^1.6.0"
shapely = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-mock = "^3.12.0"
requests-mock = "^1.11.0"
deptry = "^0.16.2"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"
ipykernel = "^7.0.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.2.7"
mkdocstrings = {extras = ["python"], version = "^0.26.1"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["simplesoilprofile"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"



[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true



[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
